feat: config,logging,exception

+++ src/events/events.controller.ts
@@ -4,6 +4,8 @@ import {
   Delete,
   Get,
   HttpCode,
+  Logger,
+  NotFoundException,
   Param,
   ParseIntPipe,
   Patch,
@@ -18,11 +20,16 @@ import { Event } from './entities/event.entity'
 
 @Controller('events')
 export class EventsController {
+  private readonly logger = new Logger(EventsController.name)
+
   constructor(@InjectRepository(Event) private readonly repository: Repository<Event>) {}
 
   @Get()
   async findAll() {
-    return await this.repository.find()
+    this.logger.log('Hit the findAll route')
+    const events = await this.repository.find()
+    this.logger.debug(`Found ${events.length} events`)
+    return events
   }
 
   @Get('practice')
@@ -47,12 +54,15 @@ export class EventsController {
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id) {
-    console.log(typeof id)
-    return await this.repository.findOne({
+    const event = await this.repository.findOne({
       where: {
         id,
       },
     })
+    if (!event) {
+      throw new NotFoundException()
+    }
+    return event
   }
 
   @Post()
@@ -70,6 +80,11 @@ export class EventsController {
         id,
       },
     })
+
+    if (!event) {
+      throw new NotFoundException()
+    }
+
     return await this.repository.save({
       ...event,
       ...input,
@@ -85,6 +100,11 @@ export class EventsController {
         id,
       },
     })
+
+    if (!event) {
+      throw new NotFoundException()
+    }
+
     await this.repository.remove(event)
   }
 }
